Make CodeProgress own CodeProgressHelper
This should make it easier to avoid reinitialising the OMODFramework
each time a new OMOD is installed.

A new helper needs constructing each time due to the C++/CLI insisting
that reusing an IDisposable is like reusing an object with placement
new - once Dispose is called, it calls Dispose on all IDisposable member
variables and GC.SuppressFinalize(this), so there's no opportunity to
delete anything later.