Use temp directory within mod directory
This avoids having to use the disk root, but still allows us to perform
the last step of installation as a rename rather than a copy-and-delete.

Just in case OMODFramework tries using the temp path before we've
created the mod, I've wrapped it up as a stack, so we push the
mod-specific path when it exists and pop it when it gets destroyed,
leaving the default path available the rest of the time.