Make OMODFramework aware of active BSAs
At least one OMOD needs to extract files from existing BSAs. Thankfully,
basically everything needed for that was already in OMODFramework, and
it just needed telling which BSAs were active at the time of mod
installtion. Annoyingly, it took them as a HashSet, so I needed to
change to an ordered collection, hence the version bump. Also, MO2 won't
tell plugins which BSAs are actually active, so this has to reconstruct
the list. It assumes archive invalidation is active in some form or
other, but MO2 has that built-in, and presumably no one's ever turned it
off.

This doesn't actually fix the problematic mod as it then packs a new BSA
which OMODFramework can't handle yet.